spl: Reduce the size of the bl_len field
This is a block length, so typicaly 512 bytes. Reduce the size to
16 bits to save space, before more fields are added in future work.

+++ include/spl.h
@@ -309,7 +309,7 @@ struct spl_load_info {
 	spl_load_reader read;
 	void *priv;
 #if IS_ENABLED(CONFIG_SPL_LOAD_BLOCK)
-	int bl_len;
+	u16 bl_len;
 #endif
 };
